+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### ✨ Added
 
 ### 🔄 Changed
+- Renamed `n_total` to `n_samples` in `glide.simulators` dataset generators.
 
 ### 🐛 Fixed